requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
astroid==3.0.1
attrs==23.1.0
colorama==0.4.6
dill==0.3.7
frozenlist==1.4.0
idna==3.6
isort==5.13.1
mccabe==0.7.0
multidict==6.0.4
platformdirs==4.1.0
pylint==3.0.3
python-dotenv==1.0.0
tomlkit==0.12.3
yarl==1.9.4
